organizing folder structure to subpackages
