+++ pyproject.toml
@@ -3,12 +3,12 @@ requires = ["setuptools", "setuptools-scm"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "eval_reg"
+name = "aind_registration_evaluation"
 description = "Evaluates registration."
 license = {text = "MIT"}
 requires-python = ">=3.7"
 classifiers = [
-    "Programming Language :: Python :: 3"
+  "Programming Language :: Python :: 3"
 ]
 dynamic = ["version", "readme"]
 
@@ -46,7 +46,7 @@ dev = [
 where = ["src"]
 
 [tool.setuptools.dynamic]
-version = {attr = "eval_reg.__version__"}
+version = {attr = "aind_registration_evaluation.__version__"}
 readme = {file = ["README.md"]}
 
 [tool.black]
@@ -72,7 +72,7 @@ exclude = '''
 
 [tool.coverage.run]
 omit = ["*__init__*"]
-source = ["eval_reg", "tests"]
+source = ["aind_registration_evaluation", "tests"]
 
 [tool.coverage.report]
 exclude_lines = ["if __name__ == .__main__.:"]
