+++ src/aind_registration_evaluation/__init__.py
@@ -1,0 +1,3 @@
+"""Package to evaluate image stitching
+"""
+__version__ = "0.0.5"
