+++ src/aind_registration_evaluation/_shared/__init__.py
@@ -1,0 +1,3 @@
+"""
+Shared constants
+"""
