+++ src/aind_registration_evaluation/default_config.yaml
