+++ src/aind_registration_evaluation/sample/__init__.py
@@ -1,0 +1,7 @@
+"""
+Methods to sample points in intersection
+areas
+"""
+# flake8: noqa: F403
+
+from .sample_points import *
