--- src/eval_reg/__init__.py
@@ -1,3 +1,0 @@
-"""Simple package to demo project structure.
-"""
-__version__ = "0.0.4"
